requirements.txt
requests==2.25.0
beautifulsoup4==4.9.3
python-telegram-bot==13.1
faunadb==4.0.1
python-dotenv==0.15.0
pyshorteners==1.0.1
sentry-sdk==0.19.5
